Harden ALC parser and prepare ReceiverBase for direct packet injection
Parser robustness:
- Bounds-check every advance of hdr_ptr in AlcPacket against the input
  length; cap lct_header_len; sanity-check expected_header_len and
  per-extension HEL; reject zero-length and oversized extensions.
  Malformed/truncated packets are now rejected cleanly instead of
  reading past the buffer end.
- Replace memberwise (uint16_t*)/(uint32_t*) casts with memcpy via
  read_unaligned<T>/write_unaligned helpers. Strict-aliasing-clean and
  no longer trips -fsanitize=alignment/undefined.
- Replace every `throw "literal"` (44 sites across AlcPacket,
  EncodingSymbol, File, FileDeliveryTable, IpSec) with
  `throw std::runtime_error(...)`. The catch in handle_received_packet
  is `catch (std::exception&)`, which never matched a const char*; the
  receiver was aborting on the first malformed packet.
- Add 4-byte SBN/ESI prefix check and encoding_symbol_length zero-guard
  in EncodingSymbol::from_payload.

ReceiverBase / receivers:
- Slim ReceiverBase ctor to (uint64_t tsi); drop the unused
  <boost/asio.hpp> include and the boost::system::error_code forward
  decl from the base header. _mcast_address/_mcast_port move to the
  derived Receiver and PcapReceiver classes that actually use them.
  This makes ReceiverBase reusable from a non-socket transport.
- Release _files_mutex before invoking the completion callback
  (snapshot the shared_ptr + callback under the lock, dispatch outside)
  so callbacks that re-enter the receiver cannot deadlock.
- Guard _completion_cb writes and file_list() reads with the same
  mutex. Drop the per-packet spdlog::info log line.

Type widening:
- SourceBlock::id, SourceBlock::symbols, File::_source_blocks and the
  FecTransformer create_blocks/extract_file signatures widen from
  uint16_t to uint32_t to match EncodingSymbol's uint32_t SBN/ESI.
- Bound checks change from `>` to `>=` so std::map::operator[] cannot
  silently insert an empty entry on out-of-range index.
- FileEntry::content_length and the FDT-parser content_length /
  transfer_length locals widen to uint64_t (RFC 5052 transfer-length
  is 48 bits); drop the lossy uint32 cast in ReceiverBase. tinyxml2's
  uint64_t SetAttribute overload is used in to_string.

FDT parser:
- Null-check every FirstChildElement / Attribute lookup and the return
  of doc.Parse. Wrap std::stoull so an invalid Expires attribute
  surfaces as a thrown runtime_error rather than a terminate.

EXT_FTI transmit path:
- Fix the upper-16-bits encoding of the 48-bit transfer-length:
  `(transfer_length >> 32) & 0xFFFF` for the high half and
  `transfer_length & 0xFFFFFFFF` for the low half. The previous mask
  `(transfer_length & 0x00FF0000) >> 32` was always zero.

Build:
- Bump C++ standard to C++23.
- Verified clean with -O3 Release and Debug + ASan + UBSan.

+++ include/FileDeliveryTable.h
@@ -61,7 +61,7 @@ namespace LibFlute {
       struct FileEntry {
         uint32_t toi;
         std::string content_location;
-        uint32_t content_length;
+        uint64_t content_length;
         std::string content_md5;
         std::string content_type;
         uint64_t expires;
